Initial work on splitting bat into bat and batwing
