+++ .gitignore
@@ -1,3 +1,5 @@
+target/
+**/*.rs.bk
 
-/target/
+/target
 **/*.rs.bk
