--- assets/.ignore
@@ -1,2 +1,0 @@
-syntaxes/*
-themes/*
