+++ batwing/Cargo.toml
@@ -1,0 +1,53 @@
+[package]
+authors = ["David Peter <mail@david-peter.de>"]
+categories = ["command-line-utilities"]
+description="A cat(1) clone with wings."
+homepage = "https://github.com/sharkdp/bat"
+license = "MIT/Apache-2.0"
+name = "batwing"
+readme = "README.md"
+repository = "https://github.com/sharkdp/bat"
+version = "0.9.0"
+exclude = [
+    "assets/syntaxes/*",
+    "assets/themes/*",
+]
+build = "build.rs"
+
+[features]
+cli = []
+
+[dependencies]
+atty = "0.2.2"
+ansi_term = "0.11"
+ansi_colours = "^1.0"
+console = "0.7"
+directories = "1.0"
+lazy_static = "1.2"
+wild = "2.0"
+content_inspector = "0.2.4"
+encoding = "0.2"
+shell-words = "0.1.0"
+
+[dependencies.git2]
+version = "0.8"
+default-features = false
+features = []
+
+[dependencies.syntect]
+version = "3.0.2"
+default-features = false
+features = ["parsing", "yaml-load", "dump-load", "dump-create"]
+
+[dependencies.clap]
+version = "2.32"
+default-features = false
+features = ["suggestions", "color", "wrap_help"]
+
+[dependencies.error-chain]
+version = "0.12"
+default-features = false
+features = []
+
+[build-dependencies]
+clap = "2.32"
