+++ batwing/assets/JavaDoc.sublime-syntax.patch
