+++ batwing/assets/create.sh
