+++ batwing/assets/syntaxes.bin
