+++ batwing/assets/syntaxes/Javascript (Babel).sublime-syntax
