+++ batwing/assets/syntaxes/PowerShell.sublime-syntax
