+++ batwing/assets/syntaxes/TypeScript.sublime-syntax
