+++ batwing/assets/syntaxes/show-nonprintable.sublime-syntax
