+++ batwing/src/app.rs
@@ -14,75 +14,13 @@ use console::Term;
 use ansi_term;
 
 use assets::BAT_THEME_DEFAULT;
-use config::{get_args_from_config_file, get_args_from_env_var};
+use config::{get_args_from_config_file, get_args_from_env_var, Config, PagingMode};
 use errors::*;
 use inputfile::InputFile;
 use line_range::{LineRange, LineRanges};
 use style::{OutputComponent, OutputComponents, OutputWrap};
 use syntax_mapping::SyntaxMapping;
 use util::transpose;
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub enum PagingMode {
-    Always,
-    QuitIfOneScreen,
-    Never,
-}
-
-#[derive(Clone)]
-pub struct Config<'a> {
-    /// List of files to print
-    pub files: Vec<InputFile<'a>>,
-
-    /// The explicitly configured language, if any
-    pub language: Option<&'a str>,
-
-    /// Whether or not to show/replace non-printable characters like space, tab and newline.
-    pub show_nonprintable: bool,
-
-    /// The character width of the terminal
-    pub term_width: usize,
-
-    /// The width of tab characters.
-    /// Currently, a value of 0 will cause tabs to be passed through without expanding them.
-    pub tab_width: usize,
-
-    /// Whether or not to simply loop through all input (`cat` mode)
-    pub loop_through: bool,
-
-    /// Whether or not the output should be colorized
-    pub colored_output: bool,
-
-    /// Whether or not the output terminal supports true color
-    pub true_color: bool,
-
-    /// Style elements (grid, line numbers, ...)
-    pub output_components: OutputComponents,
-
-    /// Text wrapping mode
-    pub output_wrap: OutputWrap,
-
-    /// Pager or STDOUT
-    pub paging_mode: PagingMode,
-
-    /// Specifies the lines that should be printed
-    pub line_ranges: LineRanges,
-
-    /// The syntax highlighting theme
-    pub theme: String,
-
-    /// File extension/name mappings
-    pub syntax_mapping: SyntaxMapping,
-
-    /// Command to start the pager
-    pub pager: Option<&'a str>,
-
-    /// Whether or not to use ANSI italics
-    pub use_italic_text: bool,
-
-    /// Lines to highlight
-    pub highlight_lines: Vec<usize>,
-}
 
 fn is_truecolor_terminal() -> bool {
     env::var("COLORTERM")
