+++ batwing/src/assets.rs
