+++ batwing/src/config.rs
@@ -7,6 +7,72 @@ use shell_words;
 
 use dirs::PROJECT_DIRS;
 use util::transpose;
+use inputfile::InputFile;
+use line_range::LineRanges;
+use style::{OutputComponents, OutputWrap};
+use syntax_mapping::SyntaxMapping;
+
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub enum PagingMode {
+    Always,
+    QuitIfOneScreen,
+    Never,
+}
+
+#[derive(Clone)]
+pub struct Config<'a> {
+    /// List of files to print
+    pub files: Vec<InputFile<'a>>,
+
+    /// The explicitly configured language, if any
+    pub language: Option<&'a str>,
+
+    /// Whether or not to show/replace non-printable characters like space, tab and newline.
+    pub show_nonprintable: bool,
+
+    /// The character width of the terminal
+    pub term_width: usize,
+
+    /// The width of tab characters.
+    /// Currently, a value of 0 will cause tabs to be passed through without expanding them.
+    pub tab_width: usize,
+
+    /// Whether or not to simply loop through all input (`cat` mode)
+    pub loop_through: bool,
+
+    /// Whether or not the output should be colorized
+    pub colored_output: bool,
+
+    /// Whether or not the output terminal supports true color
+    pub true_color: bool,
+
+    /// Style elements (grid, line numbers, ...)
+    pub output_components: OutputComponents,
+
+    /// Text wrapping mode
+    pub output_wrap: OutputWrap,
+
+    /// Pager or STDOUT
+    pub paging_mode: PagingMode,
+
+    /// Specifies the lines that should be printed
+    pub line_ranges: LineRanges,
+
+    /// The syntax highlighting theme
+    pub theme: String,
+
+    /// File extension/name mappings
+    pub syntax_mapping: SyntaxMapping,
+
+    /// Command to start the pager
+    pub pager: Option<&'a str>,
+
+    /// Whether or not to use ANSI italics
+    pub use_italic_text: bool,
+
+    /// Lines to highlight
+    pub highlight_lines: Vec<usize>,
+}
 
 pub fn config_file() -> PathBuf {
     env::var("BAT_CONFIG_PATH")
