+++ batwing/src/diff.rs
