+++ batwing/src/terminal.rs
