+++ batwing/src/util.rs
