--- src/dirs.rs
@@ -1,6 +1,0 @@
-use directories::ProjectDirs;
-
-lazy_static! {
-    pub static ref PROJECT_DIRS: ProjectDirs =
-        ProjectDirs::from("", "", crate_name!()).expect("Could not get home directory");
-}
